+++ expenses/apps/expenseapp/tests.py
@@ -1,5 +1,4 @@
 from http import HTTPStatus
-from re import T
 from django.test import TestCase
 from django.contrib.auth.models import User
 from django.core.files.uploadedfile import SimpleUploadedFile